requirements.txt
# requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
email-validator
